Don't overwrite PSD once it is created.
This allows the user to work in Photoshop and use all the features of Photoshop.
The PSD will be generated, if it doesn't exist. After that the PSD will be
reused and changes with Storyboarder's tools will be ignored.
This has the advantage, that information about layers or vector graphics are
retained.
The behaviour can be configured in the Preference dialog.

+++ src/js/window/main-window.js
@@ -70,7 +70,7 @@ let layerStatus = {
   main:       { dirty: false },
   reference:  { dirty: false },
   notes:      { dirty: false },
-  
+
   composite:  { dirty: false } // TODO do we need this?
 }
 let imageFileDirtyTimer
@@ -206,7 +206,7 @@ const commentOnLineMileage = (miles) => {
       ]
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       break
-    case 1: 
+    case 1:
       otherMessages = [
         "Looking great!!!",
         "Absolutely fantastic!",
@@ -222,7 +222,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.playEffect('tool-pencil')
       break
-    case 5: 
+    case 5:
       message.push('5 line miles.')
       otherMessages = [
         "You should be done with your rough drawing.",
@@ -238,7 +238,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.playEffect('tool-light-pencil')
       break
-    case 8: 
+    case 8:
       message.push('8 line miles.')
       otherMessages = [
         "Let's finish this up!",
@@ -254,7 +254,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.playEffect('tool-brush')
       break
-    case 10: 
+    case 10:
       message.push('10 miles!')
       otherMessages = [
         "Let's finish this up!",
@@ -270,7 +270,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.positive()
       break
-    case 20: 
+    case 20:
       message.push('20 miles!!!')
       otherMessages = [
         "This is done. Let's move on.",
@@ -285,7 +285,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.negative()
       break
-    case 50: 
+    case 50:
       message.push('50 miles!!!')
       otherMessages = [
         "Uhh.. I fell asleep. What did I miss?",
@@ -300,7 +300,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.negative()
       break
-    case 100: 
+    case 100:
       message.push('100 miles!!!')
       otherMessages = [
         "Nope!!! I'm going to delete this board if you keep drawing. Just kidding. Or am I?",
@@ -316,7 +316,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.error()
       break
-    case 200: 
+    case 200:
       message.push('200 miles!!!')
       otherMessages = [
         "Now you're just fucking with me.",
@@ -327,7 +327,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.error()
       break
-    case 300: 
+    case 300:
       message.push('300 miles!!!')
       otherMessages = [
         "I quit.",
@@ -338,7 +338,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.error()
       break
-    case 500: 
+    case 500:
       message.push('500 miles!!!')
       otherMessages = [
         "So close to 1000!!!",
@@ -346,7 +346,7 @@ const commentOnLineMileage = (miles) => {
       message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
       sfx.error()
       break
-    case 1000: 
+    case 1000:
       message.push('1000 miles!!!')
       otherMessages = [
         "Great job. :/ See ya.",
@@ -376,7 +376,7 @@ let loadBoardUI = ()=> {
     document.getElementById('storyboarder-sketch-pane'),
     size
   )
-  
+
   window.addEventListener('resize', () => {
     resize()
     setTimeout(() => storyboarderSketchPane.resize(), 500) // TODO hack, remove this #440
@@ -538,8 +538,8 @@ let loadBoardUI = ()=> {
     })
   }
 
-  
-    
+
+
     // for (var item of document.querySelectorAll('.thumbnail')) {
     //   item.classList.remove('active')
     // }
@@ -593,7 +593,7 @@ let loadBoardUI = ()=> {
       if (el) {
         offset = el.getBoundingClientRect().width
         el = thumbnailFromPoint(x, y, offset/2)
-      } 
+      }
 
       if (!el) {
         console.warn("couldn't find nearest thumbnail")
@@ -710,7 +710,7 @@ let loadBoardUI = ()=> {
     }
     sfx.playEffect('metal')
   })
-  
+
   toolbar.on('grid', value => {
     guides.setState({ grid: value })
     sfx.playEffect('metal')
@@ -1048,7 +1048,7 @@ let insertNewBoardsWithFiles = (filepaths) => {
           // thumbnail
           const thumbnailHeight = 60
           let thumbRatio = thumbnailHeight / boardSize.height
-          
+
           image.width = (image.width / boardSize.width) * (thumbRatio * boardSize.width)
           image.height = image.height / boardSize.height * 60
           canvas.width = thumbRatio * boardSize.width
@@ -1118,8 +1118,8 @@ let markImageFileDirty = layerIndices => {
 
 const addToLineMileage = value => {
   let board = boardData.boards[currentBoard]
-  if (!(board.lineMileage)) { 
-    board.lineMileage = 0 
+  if (!(board.lineMileage)) {
+    board.lineMileage = 0
   }
   let mileageChecks = [0.01,1,5,8,10,20,50,100,200,300,1000]
   for (let checkAmount of mileageChecks) {
@@ -1213,7 +1213,7 @@ let saveImageFile = () => {
       }
     }
   }
-  
+
   if (shouldSaveBoardFile) {
     saveBoardFile()
   }
@@ -1229,37 +1229,41 @@ let saveImageFile = () => {
 }
 
 let openInEditor = () => {
-    let children = ['reference', 'main', 'notes'].map((layerName, i) => {
-      return {
-        "id": (i+2),
-        "name": layerName,
-        "canvas": storyboarderSketchPane.getLayerCanvasByName(layerName)
-      }
-    });
-    var whiteBG = document.createElement('canvas')
-    whiteBG.width = storyboarderSketchPane.canvasSize[0]
-    whiteBG.height = storyboarderSketchPane.canvasSize[1]
-    var whiteBGContext = whiteBG.getContext('2d')
-    whiteBGContext.fillStyle = 'white'
-    whiteBGContext.fillRect(0, 0, whiteBG.width, whiteBG.height)
-    children = [{
-      "id": 1,
-      "name": "Background",
-      "canvas": whiteBG
-    }].concat(children)
-    let psd = {
-      width: storyboarderSketchPane.canvasSize[0],
-      height: storyboarderSketchPane.canvasSize[1],
-      imageResources: {layerSelectionIds: [3] },
-      children: children
-    };
     let board = boardData.boards[currentBoard]
     let imageFilePath = path.join(boardPath, 'images', board.url.replace('.png', '.psd'))
-    
-    console.log(psd)
 
-    const buffer = writePsd(psd);
-    fs.writeFileSync(imageFilePath, buffer);
+    if (prefsModule.getPrefs()['enablePSDdelete'] || !fs.existsSync(imageFilePath)) {
+        let children = ['reference', 'main', 'notes'].map((layerName, i) => {
+          return {
+            "id": (i+2),
+            "name": layerName,
+            "canvas": storyboarderSketchPane.getLayerCanvasByName(layerName)
+          }
+        });
+        var whiteBG = document.createElement('canvas')
+        whiteBG.width = storyboarderSketchPane.canvasSize[0]
+        whiteBG.height = storyboarderSketchPane.canvasSize[1]
+        var whiteBGContext = whiteBG.getContext('2d')
+        whiteBGContext.fillStyle = 'white'
+        whiteBGContext.fillRect(0, 0, whiteBG.width, whiteBG.height)
+        children = [{
+          "id": 1,
+          "name": "Background",
+          "canvas": whiteBG
+        }].concat(children)
+        let psd = {
+          width: storyboarderSketchPane.canvasSize[0],
+          height: storyboarderSketchPane.canvasSize[1],
+          imageResources: {layerSelectionIds: [3] },
+          children: children
+        };
+
+        console.log(psd)
+
+        const buffer = writePsd(psd);
+        fs.writeFileSync(imageFilePath, buffer);
+    }
+
     shell.openItem(imageFilePath);
 
     fs.watchFile(imageFilePath, (cur, prev) => {
@@ -1275,7 +1279,7 @@ let openInEditor = () => {
         storeUndoStateForImage(true, [0, 1, 3])
         isCurrentBoard = true
       }
-      
+
       psdData = FileHelper.getBase64ImageDataFromFilePath(imageFilePath, readerOptions)
       if(!psdData || !psdData.main) {
         return;
@@ -1317,7 +1321,7 @@ let openInEditor = () => {
 //       let imageData = canvas
 //         .toDataURL('image/png')
 //         .replace(/^data:image\/\w+;base64,/, '')
-    
+
 //       try {
 //         fs.writeFile(imageFilePath, imageData, 'base64', () => {
 //           resolve()
@@ -1370,7 +1374,7 @@ const saveThumbnailFile = (index, options = { forceReadFromFiles: false }) => {
       let imageData = canvas
         .toDataURL('image/png')
         .replace(/^data:image\/\w+;base64,/, '')
-    
+
       try {
         fs.writeFile(imageFilePath, imageData, 'base64', () => {
           console.log('saved thumbnail', imageFilePath)
@@ -1593,11 +1597,11 @@ let gotoBoard = (boardNumber, shouldPreserveSelections = false) => {
     currentBoard = boardNumber
     currentBoard = Math.max(currentBoard, 0)
     currentBoard = Math.min(currentBoard, boardData.boards.length-1)
-    
+
     if (!shouldPreserveSelections) selections.clear()
     selections = new Set([...selections.add(currentBoard)].sort(util.compareNumbers))
     renderThumbnailDrawerSelections()
-    
+
     for (var item of document.querySelectorAll('.thumbnail')) {
       item.classList.remove('active')
     }
@@ -1746,7 +1750,7 @@ let renderMetaData = () => {
   // TODO how to regenerate tooltips?
   // if (boardData.defaultBoardTiming) {
   //   document.querySelector('input[name="duration"]').dataset.tooltipDescription = `Enter the number of milliseconds for a board. There are 1000 milliseconds in a second. ${boardData.defaultBoardTiming} milliseconds is the default.`
-  // 
+  //
   //   let defaultFramesPerBoard = Math.round(boardData.defaultBoardTiming / 1000 * 24)
   //   document.querySelector('input[name="frames"]').dataset.tooltipDescription = `Enter the number of frames for a board. There are 24 frames in a second. ${defaultFramesPerBoard} frames is the default.`
   // }
@@ -1779,11 +1783,11 @@ const renderStats = () => {
 
   let stats = []
   let totalNewShots = boardData.boards.reduce((a, b) => a + (b.newShot ? 1 : 0), 0) || 1
-  secondaryStats.push( 
+  secondaryStats.push(
     `${boardData.boards.length} ${util.pluralize(boardData.boards.length, 'board').toUpperCase()}, ` +
     `${totalNewShots} ${util.pluralize(totalNewShots, 'shot').toUpperCase()}`
   )
-  
+
   let totalLineMileage = boardData.boards.reduce((a, b) => a + (b.lineMileage || 0), 0)
   let avgLineMileage = totalLineMileage / boardData.boards.length
   secondaryStats.push( (avgLineMileage/5280).toFixed(1) + ' AVG. LINE MILEAGE' )
@@ -1801,9 +1805,9 @@ const renderStats = () => {
 
   // if (scriptData) {
   //   let numScenes = scriptData.filter(data => data.type == 'scene').length
-  
+
   //   let numBoards = 'N' // TODO sum total number of boards in the script
-  
+
   //   document.querySelector('#right-stats .stats-primary').innerHTML = `${numScenes} SCENES ${numBoards} BOARDS`
   // } else {
   //   let numBoards = boardData.boards.length
@@ -1881,7 +1885,7 @@ let updateSketchPaneBoard = () => {
   return new Promise((resolve, reject) => {
     // get current board
     let board = boardData.boards[currentBoard]
-    
+
 
     // always load the main layer
     let layersData = [
@@ -2172,7 +2176,7 @@ let renderThumbnailDrawer = ()=> {
         renderThumbnailDrawerSelections()
       } else if (currentBoard !== index) {
         // go to board by index
-        
+
         // reset selections
         selections.clear()
 
@@ -2186,7 +2190,7 @@ let renderThumbnailDrawer = ()=> {
 
   renderThumbnailButtons()
   renderTimeline()
-  
+
 
   //gotoBoard(currentBoard)
 }
@@ -2210,7 +2214,7 @@ let renderThumbnailButtons = () => {
       <div class="icon">✚</div>
     `
     drawerEl.appendChild(el)
-    
+
     el.addEventListener('pointerdown', event => {
       let eventMouseOut = document.createEvent('MouseEvents')
       eventMouseOut.initMouseEvent('mouseout', true, true)
@@ -2436,7 +2440,7 @@ let setDragTarget = (x) => {
 
   let mouseX = x - containerRect.left
   let midpointX = containerRect.width / 2
-  
+
   // distance ratio -1...0...1
   let distance = (mouseX - midpointX) / midpointX
 
@@ -2446,7 +2450,7 @@ let setDragTarget = (x) => {
   if (distance < -0.5)
   {
     strength = -util.norm(distance, -0.5, -1)
-  } 
+  }
   // 0.5..1
   else if (distance > 0.5)
   {
@@ -2469,7 +2473,7 @@ let updateDrag = () => {
     return
   }
 
-  
+
   if (isEditMode && dragMode) {
     setDragTarget(lastPointer.x)
     updateThumbnailCursor(lastPointer.x, lastPointer.y)
@@ -3557,9 +3561,9 @@ let moveSelectedBoards = (position) => {
   if (position > firstSelection) {
     position = position - numRemoved
   }
-  
-  console.log('move starting at board', firstSelection, 
-              ', moving', numRemoved, 
+
+  console.log('move starting at board', firstSelection,
+              ', moving', numRemoved,
               'boards to index', position)
 
   boardData.boards.splice(position, 0, ...movedBoards)
@@ -3680,11 +3684,11 @@ let updateThumbnailCursor = (x, y) => {
   if (el) {
     offset = el.getBoundingClientRect().width
     el = thumbnailFromPoint(x, y, offset/2)
-  } 
+  }
 
   if (el) thumbnailCursor.el = el // only update if found
   if (!el) return
-  
+
   // store a reference to the nearest thumbnail
   thumbnailCursor.el = el
 
@@ -3696,14 +3700,14 @@ let updateThumbnailCursor = (x, y) => {
                       el.offsetParent.offsetParent.scrollLeft
 
   let elementOffsetX = el.getBoundingClientRect().right
-  
+
   // is this an end shot?
   if (el.classList.contains('endShot')) {
     elementOffsetX += 5
   }
 
   let arrowOffsetX = -8
-  
+
   thumbnailCursor.x = sidebarOffsetX +
                       scrollOffsetX +
                       elementOffsetX +
@@ -3777,9 +3781,9 @@ const welcomeMessage = () => {
   ]
   message.push(otherMessages[Math.floor(Math.random()*otherMessages.length)])
   notifications.notify({message: message.join(' '), timing: 10})
-} 
+}
 
-const setupRandomizedNotifications = () => {  
+const setupRandomizedNotifications = () => {
   let defaultMessages = util.shuffle(NotificationData.messages)
   setTimeout(()=>{welcomeMessage()}, 1000)
   setTimeout(()=>{runRandomizedNotifications(defaultMessages)}, 3000)
@@ -3808,7 +3812,7 @@ const getSceneObjectByIndex = (index) =>
   scriptData && scriptData.find(data => data.type == 'scene' && data.scene_number == index + 1)
 
 const storeUndoStateForScene = (isBefore) => {
-  let scene = getSceneObjectByIndex(currentScene) 
+  let scene = getSceneObjectByIndex(currentScene)
   // sceneId is allowed to be null (for a single storyboard with no script)
   let sceneId = scene && scene.scene_id
   undoStack.addSceneData(isBefore, { sceneId : sceneId, boardData: util.stringifyClone(boardData) })
@@ -4081,16 +4085,16 @@ ipcRenderer.on('printWorksheet', (event, args) => {
 
   if (!printWindow) {
     printWindow = new remote.BrowserWindow({
-      width: 1200, 
-      height: 800, 
-      minWidth: 600, 
-      minHeight: 600, 
+      width: 1200,
+      height: 800,
+      minWidth: 600,
+      minHeight: 600,
       backgroundColor: '#333333',
-      show: false, 
-      center: true, 
-      parent: remote.getCurrentWindow(), 
-      resizable: true, 
-      frame: false, 
+      show: false,
+      center: true,
+      parent: remote.getCurrentWindow(),
+      resizable: true,
+      frame: false,
       modal: true
     })
     printWindow.loadURL(`file://${__dirname}/../../print-window.html`)
@@ -4115,16 +4119,16 @@ ipcRenderer.on('importFromWorksheet', (event, args) => {
 ipcRenderer.on('importWorksheets', (event, args) => {
   if (!importWindow) {
     importWindow = new remote.BrowserWindow({
-      width: 1200, 
-      height: 800, 
-      minWidth: 600, 
-      minHeight: 600, 
+      width: 1200,
+      height: 800,
+      minWidth: 600,
+      minHeight: 600,
       backgroundColor: '#333333',
-      show: false, 
-      center: true, 
-      parent: remote.getCurrentWindow(), 
-      resizable: true, 
-      frame: false, 
+      show: false,
+      center: true,
+      parent: remote.getCurrentWindow(),
+      resizable: true,
+      frame: false,
       modal: true
     })
     importWindow.loadURL(`file://${__dirname}/../../import-window.html`)
